requirements.txt
apex==0.9.10dev
cityscapesScripts==2.2.2
dominate==2.5.1
kornia==0.4.1
matplotlib==3.2.1
mmcv==2.1.0
mmcv_full==1.5.0
mmseg==1.3.0
mmsegmentation.egg==info
numpy==1.19.5
opencv_contrib_python==4.5.2.52
packaging==21.3
Pillow==10.2.0
prettytable==2.5.0
pycuda==2022.1
scikit_learn==1.4.0
scipy==1.4.1
sets==0.3.2
tensorflow==2.5.0
tensorrt==8.5.1.7
timm==0.5.4
torch==1.8.0+cu111
torchvision==0.9.0+cu111
